Update merchants collection
